backend/pyproject.toml
[project]
name = "finance-dashboard"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.12",
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
